pyproject.toml
[project]
name = "web-api"
version = "0.1.0"
description = ""
authors = [
    { name = "Ömer Asaf Karasu", email = "omer.karasu@niafix.com" }
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "hezarfenai (>=0.1.1,<0.2.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
